Just a small Trivia Quiz Change
You had listed A as "All of the above" when A was specified as "(Secure) Hash". Not trying to be rude or anything; I think what you're trying to teach is pretty cool & relevant. So I simply added "All of the above" as a choice then listed the correct choice with the answer.